requirements.txt
asgiref==3.5.2
certifi==2022.6.15.1
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
Django==4.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
idna==3.3
music-tag==0.4.3
mutagen==1.45.1
Pillow==9.2.0
pycparser==2.21
PyJWT==2.4.0
pytz==2022.2.1
requests==2.28.1
sqlparse==0.4.2
urllib3==1.26.12
